requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
Jinja2==3.1.4
